scripts/requirements.txt
absl-py==0.9.0
ann-visualizer==2.5
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
astunparse==1.6.3
audiomentations==0.13.0
audioread==2.1.8
autokeras==1.0.3
autopep8==1.5.4
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudstorage==0.11.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dill==0.3.2
distlib==0.3.0
efficientnet==1.1.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
gcloud==0.18.3
google-api-core==1.21.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphviz==0.16
grpcio==1.30.0
gviz-api==1.9.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
image-classifiers==1.0.0b1
imageio==2.9.0
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
jupyter-client==6.1.7
jupyter-core==4.6.3
kapre==0.3.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-mixnets==0.1.0
Keras-Preprocessing==1.1.2
keras-tuner @ git+https://github.com/keras-team/keras-tuner.git@b9f314cde07ec6d54fdf5c7db1cd8f2c76ac53e6
kiwisolver==1.1.0
librosa==0.8.0
llvmlite==0.31.0
Markdown==3.2.2
matplotlib==3.1.3
networkx==2.4
nlpaug==1.1.0
numba==0.48.0
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pooch==1.1.1
prompt-toolkit==3.0.7
protobuf==3.12.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.6.0
pycparser==2.19
pydot==1.2.3
pydotplus==2.0.2
Pygments==2.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-magic==0.4.18
python-speech-features==0.6
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.22.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
SpecAugment==1.2.5
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-profile==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.10.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tifffile==2020.7.4
toml==0.10.1
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typeguard==2.9.1
urllib3==1.25.9
virtualenv==20.0.21
Wave==0.0.2
wcwidth==0.1.8
webapp2==2.5.2
Werkzeug==1.0.1
wrapt==1.12.1
xlwt==1.3.0
zipp==3.1.0
